images/archlinux-base: allow building on debian hosts
Unfortunately, devtmpfs is a singleton pseudo-filesystem. Meaning, even
if we mount it in containers, it will match the host-filesystem. Hence,
we might share some characteristics with our host.

In one case, debian seems to symlink /dev/shm to /run/shm, as such we
cannot mount 'over' the symlink, unless its target exists. Hence,
always create the empty directory /run/shm to prepare for this.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>