images: prepare fedora cross-arch CI images
This introduces the fedora-ci-base image, which serves as CI base image
for our non-native architecture images. The fedora-ci-via-i686 is the
first example non-native architecture that uses this base.

The underlying idea here is to get Docker images that can run natively
on our CI infrastructure (which usually means x86-64), but the image
contains a non-native Fedora in '/ci/sysroot' which we can then execute
via qemu inside of the docker container. This way we get non-native VMs
on any CI that can run Docker.

For now, this just contains the preparation for the images. The actual
qemu invocation will be added in follow ups.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>